butcher workflow ts
Break task loading methods out into 3 methods (load, set, change). Remove location.go and change to router navigations. Add a condition where you can forcibly load into a task on init, if the task ID is the parameter. (i.E. you pressed the back button and now we want to load a task from the URL). If you try to go to a task in a bad state, load back the last task you were on. If you started in a bad state and continued on to a bad state, you're SOL